requirements.txt
SpeechRecognition>=3.8.1
google-cloud-speech>=2.11.0
google-cloud-translate>=3.7.0
gtts>=2.2.3
PyAudio>=0.2.11
pydantic>=1.8.2
python-dotenv>=0.19.0

# Audio Processing
numpy>=1.21.2
sounddevice>=0.4.3
wave>=0.0.2
ffmpeg-python>=0.2.0

# Utility
requests>=2.26.0
typing-extensions>=3.10.0
python-json-logger>=2.0.2
tenacity>=8.0.1

# Development
pytest>=6.2.5
black>=21.9b0
mypy>=0.910
pylint>=2.11.1

# Type Stubs
types-requests>=2.25.9
types-PyYAML>=5.4.10
types-python-dateutil>=2.8.3

# Optional
jupyterlab>=3.2.0  # For development and testing
matplotlib>=3.4.3